+++ src/main/kotlin/crimera/patches/twitter/timeline/showpollresults/fingerprints/JsonCardInstanceDataFingerprint.kt
@@ -1,0 +1,9 @@
+package crimera.patches.twitter.timeline.showpollresults.fingerprints
+
+import app.revanced.patcher.fingerprint.MethodFingerprint
+
+object JsonCardInstanceDataFingerprint: MethodFingerprint(
+    strings = listOf(
+        "binding_values"
+    )
+)